app/backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# AI/ML
openai==1.59.6
httpx==0.28.1

# Image Processing
opencv-python==4.10.0.84
numpy==2.2.5

# Cloud Storage (CloudFlare R2 - S3 compatible)
boto3==1.35.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
qrcode[pil]==8.2
redis==5.2.1
requests==2.32.3
pillow==11.1.0

# Automation (Daily DB Expansion)
apscheduler==3.10.4
beautifulsoup4==4.12.3
lxml==5.3.0
sshtunnel==0.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
aiosqlite==0.20.0
